Refactor: re-organize packages and introduce SQL base trait for AST
This creates `ast` and `parser` sub-packages and introduces `SQL` as a
base trait for all the AST types, so that they get a `toSQL` method
that can be used in a special string context for printing the AST back
to raw SQL. A special string context is used rather than just
implementing toString because the default toString for a case class is
very useful when debugging. A consequence of this is that any SQL
object names that were just strings before are now an Identifier.